requirements.txt
streamlit>=1.30
tensorflow-cpu==2.12.1
numpy==1.23.5
pillow==9.5.0